Update where types live
